analytics-service/requirements.txt
flask
confluent_kafka
grpcio
protobuf
requests
